requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
Flask-MySQLdb==1.0.1
pymysql==1.1.0